Implement real-time sharing based on Automerge.
This introduces some architectural changes:

There is no longer a HTTPTodoList, as HTTP is no longer used to
transport TODOs, but Automerge CRDT messages.

The DataBaseTodoList has also been removed, as its API no longer
made sense in the context of collaborative editing. It may be
reintroduced later.

MemoryTodoList is renamed to TodoList and is now part of the
application. There is one instance in each client, and one in the
server. Automerge takes care of synchronising them all.

The Cucumber step definitions have parameter types for the
context/action/outcome todo lists.

The scenarios have been changed to introduce people, allowing us
to express the collaborative nature of the application.